refactor(session): extract shared session parsing out of analysis module
Both `analysis` (tool-call metrics) and `usage` (token counts) consume the
same parsed `CodeAnalysis` shape, yet the parser lived under `analysis` and
`usage::calculator` reached across into it. That's backwards: parsing is a
shared concern, not an implementation detail of either feature.

Move the parser, detector, per-provider analyzers, and common state into a
new `src/session/` module. `analysis` now only contains the aggregator
(`batch_analyzer`), and `usage::calculator` imports from `session` instead
of from `analysis`, so neither feature depends on the other anymore.

No symbol or behavior changes in this commit — every type, function, and
re-export keeps its current name so downstream changes can be reviewed
separately.
