+++ benches/benchmarks.rs
@@ -96,7 +96,7 @@ fn benchmark_file_parsing(c: &mut Criterion) {
             group.bench_with_input(
                 BenchmarkId::new("analyze_jsonl_file", name),
                 &path_buf,
-                |b, p| b.iter(|| vibe_coding_tracker::analysis::analyze_jsonl_file(black_box(p))),
+                |b, p| b.iter(|| vibe_coding_tracker::session::analyze_jsonl_file(black_box(p))),
             );
         }
     }
@@ -108,7 +108,7 @@ fn benchmark_file_parsing(c: &mut Criterion) {
 
 fn benchmark_format_detection(c: &mut Criterion) {
     use serde_json::json;
-    use vibe_coding_tracker::analysis::detector::detect_extension_type;
+    use vibe_coding_tracker::session::detector::detect_extension_type;
 
     let claude_data = vec![
         json!({"parentUuid": null, "type": "user", "message": {"role": "user"}}),
@@ -294,7 +294,7 @@ fn benchmark_batch_analysis(c: &mut Criterion) {
 
             // Simulate batch processing
             for (path, _name) in paths {
-                let _ = vibe_coding_tracker::analysis::analyze_jsonl_file(black_box(&path));
+                let _ = vibe_coding_tracker::session::analyze_jsonl_file(black_box(&path));
             }
         })
     });
