+++ src/cache/file_cache.rs
@@ -1,9 +1,9 @@
-use crate::analysis::AnalysisMode;
 use crate::constants::capacity;
 use crate::models::{
     CodeAnalysis, CodeAnalysisApplyDiffDetail, CodeAnalysisReadDetail, CodeAnalysisRecord,
     CodeAnalysisRunCommandDetail, CodeAnalysisWriteDetail, ExtensionType,
 };
+use crate::session::AnalysisMode;
 use anyhow::Result;
 use lru::LruCache;
 use std::fs;
@@ -113,8 +113,8 @@ impl FileParseCache {
         // Cache miss or outdated - need to parse.
         log::debug!("LRU cache miss for {}, parsing...", path.display());
         let analysis = match provider {
-            Some(p) => crate::analysis::analyze_session_file_typed_as(path, p, AnalysisMode::Full)?,
-            None => crate::analysis::analyze_jsonl_file_typed(path)?,
+            Some(p) => crate::session::analyze_session_file_typed_as(path, p, AnalysisMode::Full)?,
+            None => crate::session::analyze_jsonl_file_typed(path)?,
         };
         let arc_analysis = Arc::new(analysis);
         let size_bytes = estimate_analysis_bytes(arc_analysis.as_ref());
