+++ src/session/detector.rs
